requirements.txt
langchain
python-dotenv
redis
redisvl
transformers
langchain-experimental
pyowm
tavily-python